+++ R/clonalHomeostasis.R
@@ -23,7 +23,9 @@
 #' in the data. 
 #' @param chain indicate if both or a specific chain should be used - 
 #' e.g. "both", "TRA", "TRG", "IGH", "IGL".
-#' @param group.by The variable to use for grouping.
+#' @param group.by The variable to use for grouping
+#' @param order.by A vector of specific plotting order or "alphanumeric"
+#' to plot groups in order
 #' @param exportTable Exports a table of the data into the global 
 #' environment in addition to the visualization.
 #' @param palette Colors to use in visualization - input any 
@@ -40,6 +42,7 @@ clonalHomeostasis <- function(input.data,
                               cloneCall = "strict", 
                               chain = "both", 
                               group.by = NULL,
+                              order.by = NULL,
                               exportTable = FALSE, 
                               palette = "inferno") {
     cloneSize <- c(None = 0, cloneSize)
@@ -75,6 +78,14 @@ clonalHomeostasis <- function(input.data,
     
     #Plotting
     mat_melt <- melt(mat)
+    
+    if(!is.null(order.by)) {
+      mat_melt <- .ordering.function(vector = order.by,
+                                     group.by = "Var1", 
+                                     mat_melt)
+    }
+    
+    
     col <- length(unique(mat_melt$Var2))
     plot <- ggplot(mat_melt, aes(x=as.factor(Var1), y=value, fill=Var2)) +
         geom_bar(stat = "identity", position="fill",