+++ R/clonalQuant.R
@@ -9,21 +9,25 @@
 #' combined <- combineTCR(contig_list, 
 #'                         samples = c("P17B", "P17L", "P18B", "P18L", 
 #'                                     "P19B","P19L", "P20B", "P20L"))
-#' clonalQuant(combined, cloneCall="strict", scale = TRUE)
+#' clonalQuant(combined, 
+#'             cloneCall="strict", 
+#'             scale = TRUE)
 #'
 #' @param input.data The product of \code{\link{combineTCR}}, 
 #' \code{\link{combineBCR}}, or \code{\link{combineExpression}}.
 #' @param cloneCall How to call the clone - VDJC gene (\strong{gene}), 
 #' CDR3 nucleotide (\strong{nt}), CDR3 amino acid (\strong{aa}),
 #' VDJC gene + CDR3 nucleotide (\strong{strict}) or a custom variable 
-#' in the data. 
+#' in the data
 #' @param chain indicate if both or a specific chain should be used - 
 #' e.g. "both", "TRA", "TRG", "IGH", "IGL"
-#' @param group.by The column header used for grouping.
-#' @param scale Converts the graphs into percentage of unique clones.
-#' @param exportTable Returns the data frame used for forming the graph.
+#' @param group.by The column header used for grouping
+#' @param order.by A vector of specific plotting order or "alphanumeric"
+#' to plot groups in order
+#' @param scale Converts the graphs into percentage of unique clones
+#' @param exportTable Returns the data frame used for forming the graph
 #' @param palette Colors to use in visualization - input any 
-#' \link[grDevices]{hcl.pals}.
+#' \link[grDevices]{hcl.pals}
 #' @import ggplot2
 #' @export
 #' @concept Visualizing_Clones
@@ -32,7 +36,8 @@ clonalQuant <- function(input.data,
                         cloneCall = "strict", 
                         chain = "both", 
                         scale=FALSE, 
-                        group.by = NULL, 
+                        group.by = NULL,
+                        order.by = NULL,
                         exportTable = FALSE, 
                         palette = "inferno") {
   
@@ -94,7 +99,12 @@ clonalQuant <- function(input.data,
   if(!is.null(group.by)) {
     col <- length(unique(mat[,group.by]))
   }
-  mat[,x] = factor(mat[,x], levels = names(input.data))
+  
+  if(!is.null(order.by)) {
+    mat <- .ordering.function(vector = order.by,
+                              group.by = "values", 
+                              mat)
+  }
   
   #Plotting
   plot <- ggplot(data = mat,