+++ tests/testthat/test-clonalHomeostasis.R
@@ -9,6 +9,14 @@ test_that("clonalDiversity works", {
     clonalHomeostasis(combined, 
                     cloneCall = "gene")
   )
+  
+  expect_doppelganger(
+    "clonalHomeostasis_order_plot",
+    clonalHomeostasis(combined, 
+                      cloneCall = "gene",
+                      order.by = c("P17B","P18B","P19B","P20B","P17L","P18L","P19L","P20L"))
+  )
+  
   expect_equal(
     clonalHomeostasis(combined, 
                       cloneCall = "gene",